[contextkit] workaround: old data is returned for opened file
opened file is polled and on change notification is reread but old data
is returned almost all the time except when file size is changed. It
looks like content is read from the kernel cache which is not
renewed (while fuse getattr is called for the file and it returns new
a/c/mtime). The workaround is to open (and just close soon) other handle
for this file, in this case it is reread from fuse server. Workaround
can be removed when root cause will be found.

Signed-off-by: Denis Zalevskiy <denis.zalevskiy@jolla.com>