[metafuse] fixed race condition with poll handle. Fixes JB#9686
FileHandle is used w/o locking because locking is performed on higher
level. But notify() is called from client code and it should acquire
lock itself because fuse_notify_poll() call should not be locked - it
result in a deadlock. So, notify() accepts lock from caller and uses it
apropriately

Signed-off-by: Denis Zalevskiy <denis.zalevskiy@jolla.com>